Plenty of UI workings! (in description)
There are now 2 buttons that actually do something in the main menu: the new game button, which starts you off outside the apartment, and the credits button. There is a third button (quit) that could work, but have yet to verify that in a build.
In the apartment scene, there is a missing person canvas that will be disabled by default. When the functionality is there, it will appear with the respective name, date last seen, and a description of their disappearance or something.
Both the credits screen and the pop-up canvas contain placeholder text.